Rename encoder decoder trainer to encoder decoder operator since evaluate and predict methods were added
